[MC-1506]Put the logging action into a single file
